requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
mysqlclient==2.0.3
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
pycryptodome==3.10.4
python-decouple==3.5
pytz==2020.1
qrcode==7.3.1
requests==2.27.1
sqlparse==0.3.1
urllib3==1.26.8
whitenoise==5.2.0
